Image-Compressor/requirements.txt
pillow
opencv-python
optimize-images